Feat/solidity 0.8.26 (#81)
* chore: Update solc version to 0.8.26

* chore: Update solc version to 0.8.26

* chore: Update solc version to 0.8.26

* chore: Update solc version to 0.8.26

* refactor: Change external functions to internal in TestNexusERC721Integration contract

* refactor: Change function visibility from external to internal in TestHelper contract

* chore: Update npm dependencies to latest versions

* chore: Update solidity version to 0.8.26 with viaIR optimization

* chore: Update solidity version to 0.8.26 with viaIR optimization

* lint fix

* refactor: Update coverage script to use --ir-minimum flag

* refactor: Update coverage script to use --ir-minimum flag

* refactor: Update coverage script to use --ir-minimum flag

* chore: Update CI workflow to include git pull before committing gas report changes

* chore: Update CI workflow to include git pull before committing gas report changes

* refactor: Update coverage script to remove --ir-minimum flag

* chore: Update CI workflow to work on pr only

* feat: Update ModuleWhitelistFactory to use isModuleWhitelisted instead of isWhitelisted

The commit updates the ModuleWhitelistFactory contract to use the isModuleWhitelisted function instead of the deprecated isWhitelisted function. This change ensures that all modules are properly whitelisted before execution.

* push new gas report

* feat: Update solidity version to 0.8.26 in AbstractNexusFactory.sol

* Update solidity version to 0.8.26 in AbstractNexusFactory.sol

* feat: Update solidity version to 0.8.26 in BootstrapLib.sol

* feat: Add error messages for new events in EventsAndErrors.sol

* feat: Update TestBiconomyMetaFactory_Deployments contract to handle revert message change

* refactor: Update NexusAccountFactory contract for improved readability and maintainability

* refactor: Restrict fallback handler manipulation in ModuleManager.sol

The commit restricts the manipulation of the fallback handler in the ModuleManager contract to ensure the account's security and integrity. It prevents unauthorized uninstallation and reinstallation of the validator module, which could compromise the account's security. Additionally, it checks for the presence of an existing fallback handler to avoid overwriting it and causing unexpected behavior.

* feat: Add error handling for invalid EntryPoint address

The commit adds error handling for an invalid EntryPoint address in the Stakeable contract. This ensures that only valid addresses can be used when adding, unlocking, or withdrawing stakes.

* refactor: Update AbstractNexusFactory constructor to use require instead of revert

The commit updates the constructor of the AbstractNexusFactory contract to use the require statement instead of revert. This change improves code readability and follows best practices for error handling.

* refactor: Update BiconomyMetaFactory constructor to use require instead of revert

The commit updates the constructor of the BiconomyMetaFactory contract to use the require statement instead of revert. This change improves code readability and follows best practices for error handling.

* refactor: Use require instead of revert in ModuleWhitelistFactory constructor and addModuleToWhitelist function

The commit updates the constructor of the ModuleWhitelistFactory contract to use the require statement instead of revert. This change improves code readability and follows best practices for error handling.

It also updates the addModuleToWhitelist function to use require instead of revert. This ensures that the module address being added to the whitelist is not a zero address.

Refactor the ModuleWhitelistFactory contract to improve code readability and maintainability.

* refactor: Update Stakeable contract error handling for invalid EntryPoint address

* remove natspec

* refactor: Add InnerCallFailed error to INexusEventsAndErrors.sol

The commit adds the InnerCallFailed error to the INexusEventsAndErrors.sol interface. This error is thrown when an inner call fails. This addition enhances the error handling capabilities of the interface.

Refactor the INexusEventsAndErrors.sol interface to include the InnerCallFailed error.

* refactor: Use require instead of revert in K1ValidatorFactory constructor

The commit updates the constructor of the K1ValidatorFactory contract to use the require statement instead of revert. This change improves code readability and follows best practices for error handling.

* Update  GAS_REPORT.md and gas_report.json

* lint fix

* lint fix

* refactor: Update Stakeable contract error handling for invalid EntryPoint address